src/index.ts: move server over to typescript;
