+++ package.json
@@ -4,7 +4,8 @@
   "description": "DX for Government",
   "main": "server.mjs",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start:dev": "nodemon"
   },
   "repository": {
     "type": "git",
@@ -24,5 +25,13 @@
   "dependencies": {
     "express": "^4.17.1",
     "nunjucks": "^3.2.2"
+  },
+  "devDependencies": {
+    "@types/express": "^4.17.8",
+    "@types/node": "^14.11.8",
+    "@types/nunjucks": "^3.1.3",
+    "nodemon": "^2.0.5",
+    "ts-node": "^9.0.0",
+    "typescript": "^4.0.3"
   }
 }
